servicio_empresas/requirements.txt
Flask==3.1.2
python-dotenv==1.2.1
motor==3.7.1
Flask-JWT-Extended==4.7.1
werkzeug==3.1.3
httpx==0.28.1
flask[async]
pydantic==2.10.6
pydantic[email]
hypercorn==0.17.1
aiofiles==24.1.0
dnspython==2.7.0